+++ src/graphics/vulkan/resources/mesh/VulkanMeshManager.h
@@ -61,7 +61,7 @@ private:
     bool createVertexBuffer(std::string& errorMessage);
     bool createIndexBuffer(std::string& errorMessage);
 
-    static void computeSmoothNormals(std::vector<Vertex>& vertices, const std::vector<uint32_t>& indices);
+    static void computeSmoothNormals(std::vector<Vertex>& vertices, const std::vector<uint16_t>& indices);
 };
 
 #endif // NOBLEENGINE_VULKANMESHMANAGER_H
